requirements.txt
pandas
numpy
tensorflow
tensorflow-hub
nltk
gensim
pyenchant